@@ -1,4 +1,6 @@
-﻿using Portfolio.Controllers;
+﻿using System.Collections.Generic;
+using System.Globalization;
+
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -7,83 +9,76 @@ using Microsoft.AspNetCore.Mvc.Razor;
 using Microsoft.AspNetCore.Routing.Constraints;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using System.Collections.Generic;
-using System.Globalization;
+using Microsoft.Extensions.Hosting;
 
-namespace Portfolio
-{
-    public class Startup
-    {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
+using Portfolio.Controllers;
 
-        public IConfiguration Configuration { get; }
+namespace Portfolio {
+	public class Startup {
+		public Startup(IConfiguration configuration) {
+			Configuration = configuration;
+		}
 
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.Configure<CookiePolicyOptions>(options => {
-                // This lambda determines whether user consent for non-essential cookies is needed for a given request.
-                options.CheckConsentNeeded = context => true;
-                options.MinimumSameSitePolicy = SameSiteMode.None;
-            });
+		public IConfiguration Configuration { get; }
+
+		// This method gets called by the runtime. Use this method to add services to the container.
+		public void ConfigureServices(IServiceCollection services) {
+			services.Configure<CookiePolicyOptions>(options => {
+				// This lambda determines whether user consent for non-essential cookies is needed for a given request.
+				options.CheckConsentNeeded = context => true;
+				options.MinimumSameSitePolicy = SameSiteMode.None;
+			});
 
 
-            services.AddLocalization(options => options.ResourcesPath = "Translation")
-                .AddMvc()
-                .AddViewLocalization(LanguageViewLocationExpanderFormat.Suffix)
-                .AddDataAnnotationsLocalization();
-        }
+			services.AddLocalization(options => options.ResourcesPath = "Translation")
+				.AddMvc()
+				.AddViewLocalization(LanguageViewLocationExpanderFormat.Suffix)
+				.AddDataAnnotationsLocalization();
+		}
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
-        {
-            var supportedCultures = new List<CultureInfo> {
-                new CultureInfo("fr"),
-                new CultureInfo("fr-FR"),
-                new CultureInfo("en"),
-                new CultureInfo("en-US")
-            };
+		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env) {
+			var supportedCultures = new List<CultureInfo> {
+				new CultureInfo("fr"),
+				new CultureInfo("fr-FR"),
+				new CultureInfo("en"),
+				new CultureInfo("en-US")
+			};
 
-            var localizationOptions = new RequestLocalizationOptions
-            {
-                DefaultRequestCulture = new RequestCulture("en-US"),
-                SupportedCultures = supportedCultures,
-                SupportedUICultures = supportedCultures
-            };
-            app.UseRequestLocalization(localizationOptions);
-            localizationOptions.RequestCultureProviders.Insert(0, new UrlRequestCultureProvider());
+			var localizationOptions = new RequestLocalizationOptions {
+				DefaultRequestCulture = new RequestCulture("en-US"),
+				SupportedCultures = supportedCultures,
+				SupportedUICultures = supportedCultures
+			};
+			app.UseRequestLocalization(localizationOptions);
+			localizationOptions.RequestCultureProviders.Insert(0, new UrlRequestCultureProvider());
 
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Home/Error");
-                app.UseHsts();
-            }
+			if (env.IsDevelopment()) {
+				app.UseDeveloperExceptionPage();
+			} else {
+				app.UseExceptionHandler("/Home/Error");
+				app.UseHsts();
+			}
 
-            app.UseHttpsRedirection();
-            app.UseStaticFiles();
-            app.UseCookiePolicy();
+			app.UseHttpsRedirection();
+			app.UseStaticFiles();
+			app.UseRouting();
+			app.UseCookiePolicy();
 
-            app.UseMvc(routes => {
-                routes.MapRoute(
-                    name: "cultureRoute",
-                    template: "{culture}/{controller}/{action}/{id?}",
-                    defaults: new { controller = "Home", action = "Index" },
-                    constraints: new
-                    {
-                        culture = new RegexRouteConstraint("^[a-z]{2}(?:-[A-Z]{2})?$")
-                    });
+			app.UseEndpoints(endpoints => {
+				endpoints.MapControllerRoute(
+					name: "cultureRoute",
+					pattern: "{culture}/{controller}/{action}/{id?}",
+					defaults: new { controller = "Home", action = "Index" },
+					constraints: new {
+						culture = new RegexRouteConstraint("^[a-z]{2}(?:-[A-Z]{2})?$")
+					}
+				);
 
-                routes.MapRoute(
-                    name: "defaultHome",
-                    template: "{controller=Home}/{action=Index}/{id?}");
-            });
-        }
-    }
+				endpoints.MapControllerRoute(
+					name: "defaultHome",
+					pattern: "{controller=Home}/{action=Index}/{id?}");
+			});
+		}
+	}
 }
