@@ -1,10 +1,12 @@
-﻿using Portfolio.Models;
+﻿using System.Diagnostics;
+
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Localization;
-using System.Diagnostics;
+
+using Portfolio.Models;
 
 namespace Portfolio.Controllers {
-	public class HomeController:Controller {
+	public class HomeController : Controller {
 		public IStringLocalizer<HomeController> _localizer;
 		public HomeController(IStringLocalizer<HomeController> localizer) {
 			_localizer = localizer;
